requirements.txt
pandas
matplotlib
seaborn
sqlalchemy
scikit-learn
pymysql
cryptography
python-dotenv
